Moved code to App.js
